+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **added** a `getindex` method for a vector of layers and a vector of occurrences
 - **improved** the "Getting started" tutorial [#583]
 - **added** an overload for `explainmodel` [#582]
+- **added** a method to get a layer from a polygon [#587]
+- **added** a vignette for `reclassify`[#585]
 
 ## `v1.9.0`